pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-inject-fuzzer"
version = "0.1.0"
description = "Systematic prompt injection testing framework for LLM-integrated applications"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Justin Henderson", email = "whisk3y3@proton.me"}
]
keywords = ["security", "llm", "prompt-injection", "red-team", "fuzzing", "ai-security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "click>=8.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
semantic = [
    "sentence-transformers>=2.2",
    "torch>=2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
pif = "prompt_inject_fuzzer.cli:main"

[project.urls]
Homepage = "https://github.com/whisk3y3/prompt-inject-fuzzer"
Repository = "https://github.com/whisk3y3/prompt-inject-fuzzer"
Issues = "https://github.com/whisk3y3/prompt-inject-fuzzer/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
